Move schema into core
