+++ lib/core/schema.ts
