+++ lib/react/useQuery.ts
@@ -1,7 +1,7 @@
 import { useCallback, useMemo, useRef, useSyncExternalStore } from 'react'
 import type { QueryStatus } from '../core/figbird.js'
 import { splitConfig, type QueryConfig, type QueryDescriptor } from '../core/figbird.js'
-import { findServiceByName } from '../schema/types.js'
+import { findServiceByName } from '../core/schema.js'
 import { useFigbird } from './react.js'
 
 export interface QueryResult<T> {
